+++ aut/commands/node.py
@@ -2,7 +2,7 @@
 Implementation of `node` subcommands
 """
 
-from autcli.options import rpc_endpoint_option
+from aut.options import rpc_endpoint_option
 
 from click import group, command
 from typing import Optional
@@ -26,7 +26,7 @@ def info(rpc_endpoint: Optional[str]) -> None:
     """
     Print general information about the RPC node configuration and state.
     """
-    from autcli.utils import to_json, web3_from_endpoint_arg
+    from aut.utils import to_json, web3_from_endpoint_arg
 
     w3 = web3_from_endpoint_arg(None, rpc_endpoint)
     admin_node_info = w3.geth.admin.node_info()